transfer all api routes over to the app dir and deleted the pages dir
