+++ apps/next/app/api/route.ts
@@ -1,0 +1,3 @@
+export async function GET() {
+	return new Response('Welcome to my API')
+}
